# docs/api/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Open Archiver API",
		"version": "1.0.0",
		"description": "REST API for Open Archiver — an open-source email archiving platform. All authenticated endpoints require a Bearer JWT token obtained from `POST /v1/auth/login`, or an API key passed as a Bearer token.",
		"license": {
			"name": "SEE LICENSE IN LICENSE",
			"url": "https://github.com/LogicLabs-OU/OpenArchiver/blob/main/LICENSE"
		},
		"contact": {
			"name": "Open Archiver",
			"url": "https://openarchiver.com"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Local development"
		}
	],
	"security": [
		{
			"bearerAuth": []
		},
		{
			"apiKeyAuth": []
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "JWT obtained from `POST /v1/auth/login`. Pass as `Authorization: Bearer <token>`."
			},
			"apiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "X-API-KEY",
				"description": "API key generated via `POST /v1/api-keys`. Pass as `X-API-KEY: <key>`."
			}
		},
		"responses": {
			"Unauthorized": {
				"description": "Authentication is required or the token is invalid/expired.",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorMessage"
						},
						"example": {
							"message": "Unauthorized"
						}
					}
				}
			},
...
			"NotFound": {
				"description": "The requested resource was not found.",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorMessage"
						},
						"example": {
							"message": "Not found"
						}
					}
				}
			},
			"InternalServerError": {
				"description": "An unexpected error occurred on the server.",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorMessage"
						},
						"example": {
							"message": "Internal server error"
						}
					}
				}
			}
		},
		"schemas": {
			"ErrorMessage": {
				"type": "object",
				"properties": {
					"message": {
						"type": "string",
						"description": "Human-readable error description.",
						"example": "An error occurred."
					}
				},
				"required": ["message"]
			},
...
			"Recipient": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"nullable": true,
						"example": "John Doe"
					},
					"email": {
						"type": "string",
						"format": "email",
						"example": "john.doe@example.com"
					}
				},
				"required": ["email"]
			},
			"Attachment": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"example": "clx1y2z3a0000b4d2"
					},
					"filename": {
						"type": "string",
						"example": "invoice.pdf"
					},
					"mimeType": {
						"type": "string",
						"nullable": true,
						"example": "application/pdf"
					},
					"sizeBytes": {
						"type": "integer",
						"example": 204800
					},
					"storagePath": {
						"type": "string",
						"example": "open-archiver/attachments/abc123.pdf"
					}
				},
				"required": ["id", "filename", "sizeBytes", "storagePath"]
			},
			"ThreadEmail": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"description": "ArchivedEmail ID.",
						"example": "clx1y2z3a0000b4d2"
					},
					"subject": {
						"type": "string",
						"nullable": true,
						"example": "Re: Q4 Invoice"
					},
					"sentAt": {
						"type": "string",
						"format": "date-time"
					},
					"senderEmail": {
						"type": "string",
						"format": "email",
						"example": "finance@vendor.com"
					}
				},
				"required": ["id", "sentAt", "senderEmail"]
			},
			"ArchivedEmail": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"example": "clx1y2z3a0000b4d2"
					},
					"ingestionSourceId": {
						"type": "string",
						"example": "clx1y2z3a0000b4d2"
					},
					"userEmail": {
						"type": "string",
						"format": "email",
						"example": "user@company.com"
					},
					"messageIdHeader": {
						"type": "string",
						"nullable": true
					},
					"sentAt": {
						"type": "string",
						"format": "date-time"
					},
					"subject": {
						"type": "string",
						"nullable": true,
						"example": "Q4 Invoice"
					},
					"senderName": {
						"type": "string",
						"nullable": true,
						"example": "Finance Dept"
					},
					"senderEmail": {
						"type": "string",
						"format": "email",
						"example": "finance@vendor.com"
					},
					"recipients": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Recipient"
						}
					},
					"storagePath": {
						"type": "string"
					},
					"storageHashSha256": {
						"type": "string",
						"description": "SHA-256 hash of the raw email file, stored at archival time."
					},
					"sizeBytes": {
						"type": "integer"
					},
					"isIndexed": {
						"type": "boolean"
					},
					"hasAttachments": {
						"type": "boolean"
					},
					"isOnLegalHold": {
						"type": "boolean"
					},
					"archivedAt": {
						"type": "string",
						"format": "date-time"
					},
					"attachments": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Attachment"
						}
					},
					"thread": {
						"type": "array",
						"description": "Other emails in the same thread, ordered by sentAt. Only present on single-email GET responses.",
						"items": {
							"$ref": "#/components/schemas/ThreadEmail"
						}
					},
					"path": {
						"type": "string",
						"nullable": true
					},
					"tags": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"nullable": true
					}
				},
				"required": [
					"id",
					"ingestionSourceId",
					"userEmail",
					"sentAt",
					"senderEmail",
					"recipients",
					"storagePath",
					"storageHashSha256",
					"sizeBytes",
					"isIndexed",
					"hasAttachments",
					"isOnLegalHold",
					"archivedAt"
				]
			},
			"ArchivedEmailFeedItem": {
				"type": "object",
				"description": "Lightweight archived-email metadata. Raw EML and internal storage fields are omitted.",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"threadId": {
						"type": "string",
						"nullable": true
					},
					"ingestionSourceId": {
						"type": "string",
						"format": "uuid"
					},
					"userEmail": {
						"type": "string",
						"format": "email"
					},
					"messageIdHeader": {
						"type": "string",
						"nullable": true
					},
					"providerMessageId": {
						"type": "string",
						"nullable": true
					},
					"sentAt": {
						"type": "string",
						"format": "date-time"
					},
					"subject": {
						"type": "string",
						"nullable": true
					},
					"senderName": {
						"type": "string",
						"nullable": true
					},
					"senderEmail": {
						"type": "string",
						"format": "email"
					},
					"recipients": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Recipient"
						}
					},
					"hasAttachments": {
						"type": "boolean"
					},
					"archivedAt": {
						"type": "string",
						"format": "date-time"
					},
					"path": {
						"type": "string",
						"nullable": true
					},
					"tags": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"nullable": true
					}
				},
				"required": [
					"id",
					"threadId",
					"ingestionSourceId",
					"userEmail",
					"messageIdHeader",
					"providerMessageId",
					"sentAt",
					"subject",
					"senderName",
					"senderEmail",
					"recipients",
					"hasAttachments",
					"archivedAt",
					"path",
					"tags"
				]
			},
			"ArchivedEmailContentAddress": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string",
						"format": "email"
					}
				},
				"required": ["name", "email"]
			},
			"ArchivedEmailContentAttachment": {
				"type": "object",
				"properties": {
					"filename": {
						"type": "string",
						"nullable": true
					},
					"contentType": {
						"type": "string"
					},
					"size": {
						"type": "integer"
					},
					"contentId": {
						"type": "string",
						"nullable": true
					},
					"inline": {
						"type": "boolean"
					}
				},
				"required": ["filename", "contentType", "size", "contentId", "inline"]
			},
			"ArchivedEmailContent": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"subject": {
						"type": "string",
						"nullable": true
					},
					"sentAt": {
						"type": "string",
						"format": "date-time"
					},
					"messageId": {
						"type": "string",
						"nullable": true
					},
					"inReplyTo": {
						"type": "string",
						"nullable": true
					},
					"from": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ArchivedEmailContentAddress"
						}
					},
					"to": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ArchivedEmailContentAddress"
						}
					},
					"cc": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ArchivedEmailContentAddress"
						}
					},
					"bcc": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ArchivedEmailContentAddress"
						}
					},
					"replyTo": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ArchivedEmailContentAddress"
						}
					},
					"text": {
						"type": "string",
						"nullable": true
					},
					"html": {
						"type": "string",
						"nullable": true
					},
					"attachments": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ArchivedEmailContentAttachment"
						}
					}
				},
				"required": [
					"id",
					"subject",
					"sentAt",
					"messageId",
					"inReplyTo",
					"from",
					"to",
					"cc",
					"bcc",
					"replyTo",
					"text",
					"html",
					"attachments"
				]
			},
			"PaginatedArchivedEmails": {
				"type": "object",
				"properties": {
					"items": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ArchivedEmail"
						}
					},
					"total": {
						"type": "integer",
						"example": 1234
					},
					"page": {
						"type": "integer",
						"example": 1
					},
					"limit": {
						"type": "integer",
						"example": 10
					}
				},
				"required": ["items", "total", "page", "limit"]
			},
...
		}
	},
	"paths": {
...
		"/v1/archived-emails": {
			"get": {
				"summary": "Browse the archived email feed",
				"description": "Returns a read-only, newest-first cursor-paginated feed across archived emails the caller may read. The optional `path` parameter is an exact match, for example `ben.hoff@skan.ai/Inbox/`.",
				"operationId": "getArchivedEmailFeed",
				"tags": ["Archived Emails"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "path",
						"in": "query",
						"required": false,
						"description": "Exact archived mailbox path to include.",
						"schema": {
							"type": "string",
							"example": "ben.hoff@skan.ai/Inbox/"
						}
					},
					{
						"name": "ingestionSourceId",
						"in": "query",
						"required": false,
						"description": "Optionally restrict results to one ingestion source and its merge group.",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"description": "Number of messages to return, from 1 to 100.",
						"schema": {
							"type": "integer",
							"default": 25,
							"minimum": 1,
							"maximum": 100
						}
					},
					{
						"name": "cursor",
						"in": "query",
						"required": false,
						"description": "Opaque `nextCursor` returned by the previous page.",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Newest-first message feed.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": ["items", "nextCursor", "hasMore"],
									"properties": {
										"items": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/ArchivedEmailFeedItem"
											}
										},
										"nextCursor": {
											"type": "string",
											"nullable": true
										},
										"hasMore": {
											"type": "boolean"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid limit, path, ingestion source, or cursor.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					}
				}
			}
		},
		"/v1/archived-emails/changes": {
			"get": {
				"summary": "Poll for newly archived emails",
				"description": "Establishes or advances a durable, read-only checkpoint ordered by a transactional change log. When `cursor` is omitted, the response contains no items and returns a checkpoint at the current committed change-log position. Poll again with that cursor to receive newly archived messages in commit order, including inserts that were still in flight at initialization. A cursor must be reused by the same authenticated user with the same filters that created it.",
				"operationId": "getArchivedEmailChanges",
				"tags": ["Archived Emails"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "path",
						"in": "query",
						"required": false,
						"description": "Exact archived mailbox path to include.",
						"schema": {
							"type": "string",
							"example": "ben.hoff@skan.ai/Inbox/"
						}
					},
					{
						"name": "ingestionSourceId",
						"in": "query",
						"required": false,
						"description": "Optionally restrict results to one ingestion source and its merge group.",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"description": "Maximum number of changes to return, from 1 to 100.",
						"schema": {
							"type": "integer",
							"default": 100,
							"minimum": 1,
							"maximum": 100
						}
					},
					{
						"name": "cursor",
						"in": "query",
						"required": false,
						"description": "Opaque `nextCursor` from the previous changes response. Omit it only to initialize polling.",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Newly archived messages in change-log order and the next polling checkpoint.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": ["items", "nextCursor", "hasMore"],
									"properties": {
										"items": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/ArchivedEmailFeedItem"
											}
										},
										"nextCursor": {
											"type": "string"
										},
										"hasMore": {
											"type": "boolean"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid input, invalid cursor, or cursor filters do not match.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					}
				}
			}
		},
		"/v1/archived-emails/ingestion-source/{ingestionSourceId}": {
			"get": {
				"summary": "List archived emails for an ingestion source",
				"description": "Returns a paginated list of archived emails belonging to the specified ingestion source. Requires `read:archive` permission.",
				"operationId": "getArchivedEmails",
				"tags": ["Archived Emails"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "ingestionSourceId",
						"in": "path",
						"required": true,
						"description": "The ID of the ingestion source to retrieve emails for.",
						"schema": {
							"type": "string",
							"example": "clx1y2z3a0000b4d2"
						}
					},
					{
						"name": "page",
						"in": "query",
						"required": false,
						"description": "Page number for pagination.",
						"schema": {
							"type": "integer",
							"default": 1,
							"example": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"description": "Number of items per page.",
						"schema": {
							"type": "integer",
							"default": 10,
							"example": 10
						}
					}
				],
				"responses": {
					"200": {
						"description": "Paginated list of archived emails.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedArchivedEmails"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
		"/v1/archived-emails/{id}/content": {
			"get": {
				"summary": "Get structured archived email content",
				"description": "Parses one archived EML message and returns normalized headers, text, HTML, and attachment metadata without attachment bytes. Requires `read:archive` permission.",
				"operationId": "getArchivedEmailContentById",
				"tags": ["Archived Emails"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Parsed archived email content.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArchivedEmailContent"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					}
				}
			}
		},
		"/v1/archived-emails/{id}": {
			"get": {
				"summary": "Get a single archived email",
				"description": "Retrieves the full details of a single archived email by ID, including attachments and thread. Requires `read:archive` permission.",
				"operationId": "getArchivedEmailById",
				"tags": ["Archived Emails"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "The ID of the archived email.",
						"schema": {
							"type": "string",
							"example": "clx1y2z3a0000b4d2"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Archived email details.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArchivedEmail"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			},
			"delete": {
				"summary": "Delete an archived email",
				"description": "Permanently deletes an archived email by ID. Deletion must be enabled in system settings and the email must not be on legal hold. Requires `delete:archive` permission.",
				"operationId": "deleteArchivedEmail",
				"tags": ["Archived Emails"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "The ID of the archived email to delete.",
						"schema": {
							"type": "string",
							"example": "clx1y2z3a0000b4d2"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Email deleted successfully. No content returned."
					},
					"400": {
						"description": "Deletion is disabled in system settings, or the email is blocked by a retention policy / legal hold.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
...
	},
...
}